Add CSS variables and starting HTML structure
Start with preliminary styles to remove default margins
and padding. This will make structural styling easier.

Define variables in :root. This makes them easier to refer
to during coding.

Create the HTML structure. This will make it easier to define
the rest of the project, as there is now a framework
to base the rest of the work on.